+++ pytorch_lightning/plugins/training_type/training_type_plugin.py
@@ -28,8 +28,10 @@ from pytorch_lightning.plugins import TorchCheckpointIO
 from pytorch_lightning.plugins.io.checkpoint_plugin import CheckpointIO
 from pytorch_lightning.plugins.precision import PrecisionPlugin
 from pytorch_lightning.trainer.states import TrainerFn
+from pytorch_lightning.utilities import rank_zero_deprecation
 from pytorch_lightning.utilities.apply_func import apply_to_collection, move_data_to_device
 from pytorch_lightning.utilities.distributed import ReduceOp
+from pytorch_lightning.utilities.model_helpers import is_overridden
 from pytorch_lightning.utilities.types import _PATH
 
 TBroadcast = TypeVar("TBroadcast")
@@ -49,6 +51,11 @@ class TrainingTypePlugin(ABC):
         self.optimizers: List[Optimizer] = []
         self.lr_schedulers: List[_LRScheduler] = []
         self.optimizer_frequencies: List[int] = []
+        if is_overridden("post_dispatch", self, parent=TrainingTypePlugin):
+            rank_zero_deprecation(
+                f"`{self.__class__.__name__}.post_dispatch()` has been deprecated in v1.6 and will be removed in v1.7."
+                f" Move your implementation to `{self.__class__.__name__}.teardown()` instead."
+            )
 
     @property
     def checkpoint_io(self) -> CheckpointIO: